requirements.txt
numpy
scikit-image
matplotlib
tensorflow-gpu
keras
torch
torchvision
pandas
scipy
scikit-learn
tqdm
plotly
Pillow
ipywidgets
ipython